test: move helpers to its proper file
