--- src/helpers/test.ts
@@ -1,13 +1,0 @@
-import cn from 'classnames';
-
-// eslint-disable-next-line import/prefer-default-export
-export function renderClassName({ withClassName }: { withClassName: string }) {
-  const input = cn('class', {
-    name: true
-  });
-
-  return {
-    input,
-    expected: `${withClassName} ${input}`
-  };
-}
